+++ spec/unit/installer/installation_options_spec.rb
@@ -65,6 +65,7 @@ module Pod
           'integrate_targets' => true,
           'lock_pod_sources' => true,
           'warn_for_multiple_pod_sources' => true,
+          'warn_for_unused_master_specs_repo' => true,
           'share_schemes_for_development_pods' => false,
           'disable_input_output_paths' => false,
           'preserve_pod_file_structure' => false,